+++ src/modules/single-accommodation/components/accommodation-details.tsx
