+++ src/modules/single-accommodation/components/skeleton-details.tsx
