+++ src/modules/single-accommodation/components/skeleton-images.tsx
