+++ src/modules/single-accommodation/components/user-details.tsx
