+++ src/modules/single-accommodation/constants/icons.constant.ts
